Change the way pushstate handles form POST submissions.
Instead of just setting the state to `{ post: name=Jason&age=27&name=Anna }`
we now set the state to `{ name:["Jason","Anna"], age:["27"], __postData="name=Jason&age=27&name=Anna" }`

See the new section in the README for details.

This is a breaking change, so will result in a new major version number.
I think it's worth it though.